+++ src/three/math/Ellipsoid.js
@@ -51,7 +51,7 @@ export class Ellipsoid {
 	// returns a frame with Z indicating altitude
 	// Y pointing north
 	// X pointing east
-	constructLatLonFrame( lat, lon, target ) {
+	getEastNorthUpFrame( lat, lon, target ) {
 
 		this.getCartographicToPosition( lat, lon, 0, _pos );
 		this.getCartographicToNormal( lat, lon, _vecZ );